+++ skillMarketPrj/src/main/java/com/team4/skillmarket/community/controller/CommunityCommentWriteController.java
@@ -2,7 +2,6 @@ package com.team4.skillmarket.community.controller;
 
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.util.List;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -11,33 +10,43 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import com.google.gson.Gson;
 import com.team4.skillmarket.community.service.CommunityService;
 import com.team4.skillmarket.community.vo.CommunityCommentVo;
 import com.team4.skillmarket.member.vo.MemberVo;
 
-@WebServlet("/community/comment/list")
+
+//댓글 작성
+@WebServlet("/community/comment/write")
 public class CommunityCommentWriteController extends HttpServlet {
 
     @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 
         try {
 
-            String no = req.getParameter("no");
-            
+            HttpSession session = req.getSession();
+            MemberVo loginMember = (MemberVo) session.getAttribute("loginMember");
+            String writerNo = null;
+            if (loginMember != null) {
+                writerNo = loginMember.getMemberNo();
+                System.out.println("writerNo: " + writerNo);
+            }
+
+            String boardNo = req.getParameter("no");
+            String comment = req.getParameter("comment");
+
+            CommunityCommentVo vo = new CommunityCommentVo();
+            vo.setBoardNo(boardNo);
+            vo.setCommentContent(comment);
+            vo.setMemberNo(writerNo);
+
             CommunityService cs = new CommunityService();
-            List<CommunityCommentVo> list = cs.getCommentList(no);
-          
-            Gson gson = new Gson();
-            String jsonStr = gson.toJson(list);
+            int result = cs.commentWrite(vo);
 
-            //문자열 내보내기
-            resp.setContentType("application/json");
-            resp.setCharacterEncoding("UTF-8");
             PrintWriter out = resp.getWriter();
-            out.write(jsonStr);
-
+            if (result == 1) {
+                out.write("ok");
+            }
 
         } catch (Exception e) {
             System.out.println("글쓰기 에러");
