+++ skillMarketPrj/src/main/java/com/team4/skillmarket/community/dao/FreeBoardDao.java
@@ -19,50 +19,9 @@ import com.team4.skillmarket.community.vo.FreeBoardVo;
 
 public class FreeBoardDao {
 	
-	//게시글 한방조회용쿼리 ~ sql(WITH base_query AS 
-	//( SELECT FB.BOARD_NO, SUBSTR(FB.FREE_BOARD_TITLE, 1, 20) AS SHORT_TITLE, SUBSTR(REGEXP_REPLACE(FB.FREE_BOARD_CONTENT, '<[^>]+>', ''), 1, 15) AS SHORT_CONTENT, FB.FREE_BOARD_ENROLL_DATE, COUNT(DISTINCT FBL.LIKE_NO) OVER (PARTITION BY FB.BOARD_NO) AS LIKE_COUNT, COUNT(DISTINCT FBC.COMMENT_NO) OVER (PARTITION BY FB.BOARD_NO) AS COMMENT_COUNT, FB.FREE_BOARD_STATUS, FB.FREE_BOARD_HIT FROM FREE_BOARD FB LEFT JOIN FREE_BOARD_LIKE FBL ON FB.BOARD_NO = FBL.BOARD_NO LEFT JOIN FREE_BOARD_COMMENT FBC ON FB.BOARD_NO = FBC.BOARD_NO WHERE FB.FREE_BOARD_STATUS = 1 ORDER BY FB.FREE_BOARD_ENROLL_DATE DESC ) SELECT BOARD_NO, SHORT_TITLE, SHORT_CONTENT, LIKE_COUNT, COMMENT_COUNT, CASE WHEN EXTRACT(HOUR FROM (SYSTIMESTAMP - FREE_BOARD_ENROLL_DATE) DAY TO SECOND) < 1 THEN ROUND(EXTRACT(MINUTE FROM (SYSTIMESTAMP - FREE_BOARD_ENROLL_DATE) DAY TO SECOND)) || ' 분전' WHEN EXTRACT(DAY FROM (SYSTIMESTAMP - FREE_BOARD_ENROLL_DATE) DAY TO SECOND) < 1 THEN ROUND(EXTRACT(HOUR FROM (SYSTIMESTAMP - FREE_BOARD_ENROLL_DATE) DAY TO SECOND)) || ' 시간전' ELSE ROUND(EXTRACT(DAY FROM (SYSTIMESTAMP - FREE_BOARD_ENROLL_DATE) DAY TO SECOND)) || ' 일전' END AS TIME_AGO FROM base_query ORDER BY FREE_BOARD_STATUS ASC, FREE_BOARD_HIT DESC OFFSET 1 ROWS FETCH NEXT 20 ROWS ONLY;)
+	//게시글 한방조회용쿼리 ~ sql(SELECT FB.BOARD_NO, SUBSTR(FB.FREE_BOARD_TITLE, 1, 20) AS SHORT_TITLE, SUBSTR(REGEXP_REPLACE(FB.FREE_BOARD_CONTENT, '<[^>]+>', ''), 1, 15) AS SHORT_CONTENT, FB.FREE_BOARD_ENROLL_DATE, ( SELECT COUNT(DISTINCT FBL.LIKE_NO) FROM FREE_BOARD_LIKE FBL WHERE FBL.BOARD_NO = FB.BOARD_NO ) AS LIKE_COUNT, ( SELECT COUNT(DISTINCT FBC.COMMENT_NO) FROM FREE_BOARD_COMMENT FBC WHERE FBC.BOARD_NO = FB.BOARD_NO ) AS COMMENT_COUNT, FB.FREE_BOARD_HIT, FBCAT.FREE_BOARD_CATEGORY_NAME, CASE WHEN EXTRACT(HOUR FROM (SYSTIMESTAMP - FB.FREE_BOARD_ENROLL_DATE)) < 1 THEN ROUND(EXTRACT(MINUTE FROM (SYSTIMESTAMP - FB.FREE_BOARD_ENROLL_DATE))) || '분 전' WHEN EXTRACT(DAY FROM (SYSTIMESTAMP - FB.FREE_BOARD_ENROLL_DATE)) < 1 THEN ROUND(EXTRACT(HOUR FROM (SYSTIMESTAMP - FB.FREE_BOARD_ENROLL_DATE))) || '시간 전' WHEN EXTRACT(DAY FROM (SYSTIMESTAMP - FB.FREE_BOARD_ENROLL_DATE)) < 7 THEN ROUND(EXTRACT(DAY FROM (SYSTIMESTAMP - FB.FREE_BOARD_ENROLL_DATE))) || '일 전' ELSE EXTRACT(DAY FROM (SYSTIMESTAMP - FB.FREE_BOARD_ENROLL_DATE)) || '일 전' END AS TIME_AGO FROM FREE_BOARD FB LEFT JOIN FREE_BOARD_CATEGORY FBCAT ON FB.FREE_BOARD_CATEGORY_NO = FBCAT.FREE_BOARD_CATEGORY_NO WHERE FB.FREE_BOARD_STATUS = 1 ORDER BY FB.FREE_BOARD_ENROLL_DATE DESC OFFSET 1 ROWS FETCH NEXT 20 ROWS ONLY) 
 	public List<CommunityPostVo> getCommunityPostList(Connection conn, PageVo pv) {
-	    String sql = "WITH base_query AS (\r\n"
-	    		+ "SELECT\r\n"
-	    		+ "FB.BOARD_NO,\r\n"
-	    		+ "SUBSTR(FB.FREE_BOARD_TITLE, 1, 20) AS SHORT_TITLE,\r\n"
-	    		+ "SUBSTR(REGEXP_REPLACE(FB.FREE_BOARD_CONTENT, '<[^>]+>', ''), 1, 15) AS SHORT_CONTENT,\r\n"
-	    		+ "FB.FREE_BOARD_ENROLL_DATE,\r\n"
-	    		+ "COUNT(DISTINCT FBL.LIKE_NO) OVER (PARTITION BY FB.BOARD_NO) AS LIKE_COUNT,\r\n"
-	    		+ "COUNT(DISTINCT FBC.COMMENT_NO) OVER (PARTITION BY FB.BOARD_NO) AS COMMENT_COUNT,\r\n"
-	    		+ "FB.FREE_BOARD_STATUS,\r\n"
-	    		+ "FB.FREE_BOARD_HIT\r\n"
-	    		+ "FROM\r\n"
-	    		+ "FREE_BOARD FB\r\n"
-	    		+ "LEFT JOIN\r\n"
-	    		+ "FREE_BOARD_LIKE FBL ON FB.BOARD_NO = FBL.BOARD_NO\r\n"
-	    		+ "LEFT JOIN\r\n"
-	    		+ "FREE_BOARD_COMMENT FBC ON FB.BOARD_NO = FBC.BOARD_NO\r\n"
-	    		+ "WHERE\r\n"
-	    		+ "FB.FREE_BOARD_STATUS = 1\r\n"
-	    		+ "ORDER BY\r\n"
-	    		+ "FB.FREE_BOARD_ENROLL_DATE DESC\r\n"
-	    		+ ")\r\n"
-	    		+ "SELECT\r\n"
-	    		+ "BOARD_NO,\r\n"
-	    		+ "SHORT_TITLE,\r\n"
-	    		+ "SHORT_CONTENT,\r\n"
-	    		+ "LIKE_COUNT,\r\n"
-	    		+ "COMMENT_COUNT,\r\n"
-	    		+ "FREE_BOARD_HIT,\r\n"
-	    		+ "CASE\r\n"
-	    		+ "WHEN EXTRACT(HOUR FROM (SYSTIMESTAMP - FREE_BOARD_ENROLL_DATE) DAY TO SECOND) < 1\r\n"
-	    		+ "THEN ROUND(EXTRACT(MINUTE FROM (SYSTIMESTAMP - FREE_BOARD_ENROLL_DATE) DAY TO SECOND)) || '분전'\r\n"
-	    		+ "WHEN EXTRACT(DAY FROM (SYSTIMESTAMP - FREE_BOARD_ENROLL_DATE) DAY TO SECOND) < 1\r\n"
-	    		+ "THEN ROUND(EXTRACT(HOUR FROM (SYSTIMESTAMP - FREE_BOARD_ENROLL_DATE) DAY TO SECOND)) || '시간전'\r\n"
-	    		+ "WHEN EXTRACT(DAY FROM (SYSTIMESTAMP - FREE_BOARD_ENROLL_DATE) DAY TO SECOND) < 7\r\n"
-	    		+ "THEN ROUND(EXTRACT(DAY FROM (SYSTIMESTAMP - FREE_BOARD_ENROLL_DATE) DAY TO SECOND)) || '일전'\r\n"
-	    		+ "ELSE EXTRACT(DAY FROM (SYSTIMESTAMP - FREE_BOARD_ENROLL_DATE) DAY TO SECOND) || '일전'\r\n"
-	    		+ "END AS TIME_AGO\r\n"
-	    		+ "FROM base_query\r\n"
-	    		+ "ORDER BY\r\n"
-	    		+ "FREE_BOARD_ENROLL_DATE DESC\r\n"
-	    		+ "OFFSET ? ROWS FETCH NEXT ? ROWS ONLY";
+	    String sql = "SELECT FB.BOARD_NO, SUBSTR(FB.FREE_BOARD_TITLE, 1, 20) AS SHORT_TITLE, SUBSTR(REGEXP_REPLACE(FB.FREE_BOARD_CONTENT, '<[^>]+>', ''), 1, 15) AS SHORT_CONTENT, FB.FREE_BOARD_ENROLL_DATE, ( SELECT COUNT(DISTINCT FBL.LIKE_NO) FROM FREE_BOARD_LIKE FBL WHERE FBL.BOARD_NO = FB.BOARD_NO ) AS LIKE_COUNT, ( SELECT COUNT(DISTINCT FBC.COMMENT_NO) FROM FREE_BOARD_COMMENT FBC WHERE FBC.BOARD_NO = FB.BOARD_NO ) AS COMMENT_COUNT, FB.FREE_BOARD_HIT, FBCAT.FREE_BOARD_CATEGORY_NAME, CASE WHEN EXTRACT(HOUR FROM (SYSTIMESTAMP - FB.FREE_BOARD_ENROLL_DATE)) < 1 THEN ROUND(EXTRACT(MINUTE FROM (SYSTIMESTAMP - FB.FREE_BOARD_ENROLL_DATE))) || '분 전' WHEN EXTRACT(DAY FROM (SYSTIMESTAMP - FB.FREE_BOARD_ENROLL_DATE)) < 1 THEN ROUND(EXTRACT(HOUR FROM (SYSTIMESTAMP - FB.FREE_BOARD_ENROLL_DATE))) || '시간 전' WHEN EXTRACT(DAY FROM (SYSTIMESTAMP - FB.FREE_BOARD_ENROLL_DATE)) < 7 THEN ROUND(EXTRACT(DAY FROM (SYSTIMESTAMP - FB.FREE_BOARD_ENROLL_DATE))) || '일 전' ELSE EXTRACT(DAY FROM (SYSTIMESTAMP - FB.FREE_BOARD_ENROLL_DATE)) || '일 전' END AS TIME_AGO FROM FREE_BOARD FB LEFT JOIN FREE_BOARD_CATEGORY FBCAT ON FB.FREE_BOARD_CATEGORY_NO = FBCAT.FREE_BOARD_CATEGORY_NO WHERE FB.FREE_BOARD_STATUS = 1 ORDER BY FB.FREE_BOARD_ENROLL_DATE DESC OFFSET ? ROWS FETCH NEXT ? ROWS ONLY";
 
 	    List<CommunityPostVo> postList = new ArrayList<>();
 	    try (PreparedStatement pstmt = conn.prepareStatement(sql)) {
@@ -77,8 +36,10 @@ public class FreeBoardDao {
 	            int likeCount = rs.getInt("LIKE_COUNT");
 	            int commentCount = rs.getInt("COMMENT_COUNT");
 	            String timeAgo = rs.getString("TIME_AGO");
+	            String categoryName = rs.getString("FREE_BOARD_CATEGORY_NAME");
 	            int freeBoardHit = rs.getInt("FREE_BOARD_HIT");
-
+	            
+	            System.out.println(boardNo);
 
 	            CommunityPostVo post = new CommunityPostVo();
 	            post.setBoardNo(boardNo);
@@ -87,6 +48,7 @@ public class FreeBoardDao {
 	            post.setRecommendCount(likeCount);
 	            post.setCommentCount(commentCount);
 	            post.setFreeBoardEnrollDate(timeAgo);
+	            post.setCategoryName(categoryName);
 	            post.setFreeBoardHit(freeBoardHit);
 
 	            postList.add(post);
