+++ skillMarketPrj/src/main/webapp/static/js/estimate/estimatewrite.js
@@ -128,4 +128,3 @@ document.querySelector('#cancel-subimage-button').addEventListener('click', func
 
 
 
-
